+++ index.html
@@ -1,10 +1,10 @@
 <!DOCTYPE html>
 <html lang="ko">
     <head>
-      <link rel="stylesheet" href="css/header.css">
-      <link rel="stylesheet" href="css/navbar.css">
-      <link rel="stylesheet" href="css/common.css">
-      <link rel="stylesheet" href="css/menu.css">
+      <link rel="stylesheet" href="/css/header.css">
+      <link rel="stylesheet" href="/css/navbar.css">
+      <link rel="stylesheet" href="/css/common.css">
+      <link rel="stylesheet" href="/css/menu.css">
       <meta charset="utf-8">  
       <title>카카오페이지</title>
     </head>
@@ -268,7 +268,7 @@
         </div>
       </main>
       
-      <script type="module" src="./src/exec.js"></script>
+      <script type="module" src="/src/exec.js"></script>
       <!-- <script type="module" src="./src/main.js"></script> -->
     </body>
 </html>
